==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/00b_get_inferred_genes.ipynb <==
import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

#!pip3 install cmapBQ -q

import pandas as pd
from tqdm import tqdm
import cmapBQ.query as cmap_query
import cmapBQ.config as cmap_config

# URL with credentials
credentials_filepath='../BQ-demo-credentials.json'
# Set up credentials
cmap_config.setup_credentials(credentials_filepath)
bq_client = cmap_config.get_bq_client()

df_sig = pd.read_parquet('../../NO_BACKUP/lincs_2020/sig_overlap.parquet')

sig_ids = df_sig.sig_id.values
n = len(sig_ids)//4000



for i in tqdm(range(n + 1)):
    batch = sig_ids[i * 4000 :(i + 1) * 4000]
    cmap_query.cmap_matrix(bq_client, cid=list(batch), verbose=False, feature_space='aig').data_df.T.to_parquet('../../NO_BACKUP/lincs_2020/matrices_inferred_genes/matrix_' + str(i) + '.parquet')

#Uniting downloaded matrices to one:
i = 0
matrices = pd.read_parquet('../../NO_BACKUP/lincs_2020/matrices_inferred_genes/matrix_' + str(i) + '.parquet')
for i in tqdm(range(1, n + 1)):
    matrix = pd.read_parquet('../../NO_BACKUP/lincs_2020/matrices_inferred_genes/matrix_' + str(i) + '.parquet')
    matrices = pd.concat([matrices, matrix])

matrices.to_parquet('../../NO_BACKUP/lincs_2020/sig_matrix_filtered_inferred_genes.parquet', index=True)
