==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/01_filter_lincs_2020.ipynb <==
import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import pandas as pd

### A small filtration of a signature matrix:

siginfo_tumor_cp = pd.read_parquet('../NO_BACKUP/lincs_2020/sig_info.parquet')

siginfo_tumor_cp = siginfo_tumor_cp[(siginfo_tumor_cp['qc_pass']>0)&(siginfo_tumor_cp['pert_time']==24)]

matrix = pd.read_parquet('../NO_BACKUP/lincs_2020/sig_matrix.parquet')

len(siginfo_tumor_cp['sig_id'].values)

matrix_ = matrix[matrix.index.isin(siginfo_tumor_cp['sig_id'].values)]

matrix_.to_parquet('../NO_BACKUP/lincs_2020/sig_matrix_filtered.parquet', index=True)
