==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/03a_processing_compounds_pharmacodb.ipynb <==
import sys
import os
from tqdm import tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import json
import pandas as pd

path = '../NO_BACKUP/pharmacodb/compound_i'

len(os.listdir(path))


filename = 'compound_' + str(i) + '.json'
row = []
table = []
for i in tqdm(range(1, len(os.listdir(path)))):
    filename = 'compound_' + str(i) + '.json'
    with open(os.path.join(path, filename), 'r') as f:
        response = json.loads(f.read())
    if len(response['data']['singleCompound'])!=0:
        row.append(response['data']['singleCompound']['compound']['id'])
        row.append(response['data']['singleCompound']['compound']['name'])
        row.append(response['data']['singleCompound']['compound']['uid'])
        row.append(response['data']['singleCompound']['compound']['annotation']['inchikey'])
        row.append(response['data']['singleCompound']['compound']['annotation']['smiles'])
        row.append(response['data']['singleCompound']['compound']['annotation']['pubchem'])
        row.append(response['data']['singleCompound']['compound']['annotation']['fda_status'])
        row.append(response['data']['singleCompound']['compound']['annotation']['chembl'])
        row.append(response['data']['singleCompound']['compound']['annotation']['reactome'])
        table.append(row.copy())
        row = []


df = pd.DataFrame(table).rename(columns={0: 'id', 1: 'name', 2: 'uid', 3: 'inchikey', 4: 'smiles', 5: 'pubchem', 6: 'fda_status', 7: 'chembl', 8: 'reactome'})

df.to_parquet('../NO_BACKUP/pharmacodb/compounds_pharmdb.parqeut', index=False)
