==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/03b_processing_compounds_pharmdb.ipynb <==
#!pip3 install pandas

import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import json
import pandas as pd

with open('../NO_BACKUP/compounds.json', 'r') as f:
    comp = json.loads(f.read())

pd.DataFrame(comp)

df = pd.DataFrame(comp['data']['compounds'])

df_ann = pd.DataFrame(df['annotation'].values.tolist())

df['pubchem'] = df_ann['pubchem']
df['chembl'] = df_ann['chembl']

df = df.drop(columns=['annotation', '__typename'])

df.to_parquet('comp_info.parquet')

df.to_parquet('../NO_BACKUP/pharmacodb/comp_pharmdb.parquet', index=False)

compounds = pd.read_parquet('../NO_BACKUP/lincs_2020/comp_info.parquet')

compounds
