==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/04_processing_celllines_pharmacodb.ipynb <==
import sys
import os
from tqdm import tqdm
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')

import json
import pandas as pd

path = '../NO_BACKUP/pharmacodb/cell_line_by_id_i'

len(os.listdir(path))

row = []
table = []
for i in tqdm(range(1, len(os.listdir(path)))):
    filename = 'cell_line_by_id_' + str(i) + '.json'
    with open(os.path.join(path, filename), 'r') as f:
        response = json.loads(f.read())
        
  
    if len(response['data']['cell_line'])!=0:
        row.append(response['data']['cell_line']['id'])
        row.append(response['data']['cell_line']['name'])
        row.append(response['data']['cell_line']['uid'])
        row.append(response['data']['cell_line']['diseases'])
        row.append(response['data']['cell_line']['accession_id'])
        row.append(response['data']['cell_line']['tissue'])
        row.append(response['data']['cell_line']['synonyms'])
        table.append(row.copy())
        row = []
    

pd.DataFrame(table)

df = pd.DataFrame(table).rename(columns={0: 'id', 1: 'name', 2: 'uid', 3: 'diseases', 4: 'accession_id', 5: 'tissue', 6: 'synonyms'})

df

df.to_parquet('../NO_BACKUP/pharmacodb/cell_lines.parqeut', index=False)
