==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/07_fingerprints.ipynb <==
import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')
# https://github.com/UnixJunkie/mol2ecfp4/blob/master/ecfp4.py


#!pip3 install rdkit -q

from rdkit import Chem
from rdkit.Chem import AllChem

import pandas as pd
import numpy as np

pharm_comp = pd.read_parquet('../../NO_BACKUP/pharmacodb/compounds_pharmdb.parquet')

lincs_comp = pd.read_parquet('../../NO_BACKUP/lincs_2020/comp_info.parquet')

lincs_comp_dropna = lincs_comp[~lincs_comp['inchi_key'].isna()]

comp_overlap = lincs_comp_dropna[lincs_comp_dropna['inchi_key'].isin(pharm_comp['inchikey'].unique())]

comp_overlap = comp_overlap[~comp_overlap['pert_id'].duplicated()]

comp_overlap = comp_overlap.reset_index(drop=True)

pharm_comp[['inchikey', 'pubchem']]

comp_overlap.merge(pharm_comp[['inchikey', 'pubchem']], left_on='inchi_key', right_on='')

comp_overlap['canonical_smiles']

from rdkit.Chem import rdFingerprintGenerator
fp_diameter = 4
fp_radius = fp_diameter / 2
rdkgen = rdFingerprintGenerator.GetMorganGenerator(radius=2, fpSize=2048)
fps = []
cfps = []

for smile in comp_overlap['canonical_smiles'].values:
    if not smile is None:
        mol = Chem.MolFromSmiles(smile)
        fp = rdkgen.GetFingerprintAsNumPy(mol)
        cfp = rdkgen.GetCountFingerprintAsNumPy(mol)
        fps.append(fp)
        cfps.append(cfp)
    else:
        fps.append(None)
        cfps.append(None)

df_fps = pd.DataFrame(np.array(fps))

rename_fps_col = {}
for i in range(2048):
    rename_fps_col[i] = 'fps_' + str(i)

df_fps = df_fps.rename(columns=rename_fps_col)

df_fps['pert_id'] = comp_overlap['pert_id']

df_fps['cmap_name'] = comp_overlap['cmap_name']

df_fps['smiles'] = comp_overlap['canonical_smiles']

df_cfps = pd.DataFrame(np.array(cfps))

rename_cfps_col = {}
for i in range(2048):
    rename_cfps_col[i] = 'cfps_' + str(i)
df_cfps = df_cfps.rename(columns=rename_cfps_col)

df_cfps

df_cfps['pert_id'] = comp_overlap['pert_id']

df_cfps['cmap_name'] = comp_overlap['cmap_name']

df_cfps['smiles'] = comp_overlap['canonical_smiles']

df_fps.to_parquet('../NO_BACKUP/pharmacodb/comp_overlap_fps.parquet', index=False)
df_cfps.to_parquet('../NO_BACKUP/pharmacodb/comp_overlap_cfps.parquet', index=False)
