==> cell_line_processing/01_Lincs_PharmacoDB/notebooks/10_combine_all_the_data_inferred.ipynb <==
import sys
sys.path.insert(0,'/home/olgan96/.env/lib/python3.11/site-packages/')
import pandas as pd
import numpy as np

matrix = pd.read_parquet('../../NO_BACKUP/lincs_2020/sig_matrix_filtered_inferred_genes.parquet')

matrix[matrix.index.duplicated(keep=False)]

df_sig_resp = pd.read_parquet('../../NO_BACKUP/lincs_2020/signature_response_r2_top0.7.parquet')

df_sig_resp['pert_dose_unit'].unique()

df = pd.read_parquet('../../NO_BACKUP/tahoe/metadata/obs_metadata.parquet')

df_sig_resp[df_sig_resp['nearest_dose'] != df_sig_resp['pert_idose'].str.split().str[0].astype(float)]

df_sig_resp.loc[df_sig_resp['nearest_dose'].isna(), 'nearest_dose'] = df_sig_resp.loc[df_sig_resp['nearest_dose'].isna(), 'pert_idose'].str.split().str[0].astype(float)

df_sig_resp[df_sig_resp['nearest_dose'] != df_sig_resp['pert_idose'].str.split().str[0].astype(float)]

df_comp = pd.read_parquet('../../NO_BACKUP/pharmacodb/comp_overlap_fps.parquet')

matrix_ = matrix[matrix.index.isin(df_sig_resp['sig_id'].values)]

df_sig_resp_ = df_sig_resp[['sig_id', 'pert_id', 'cellosaurus_id', 'inchi_key', 'cell_drug', 'nearest_dose', 'responses']]

df_comp[['pert_id', 'cmap_name', 'smiles']]

df_sig_resp_ = df_sig_resp_.merge(df_comp[['pert_id', 'cmap_name', 'smiles']], on='pert_id', how='left').merge(df_comp.drop(columns=['cmap_name', 'smiles']), on='pert_id', how='left')

df_sig_resp_

df_sig_resp_final = df_sig_resp_.merge(matrix_, left_on='sig_id', right_on='cid', how='left')

df_sig_resp_final_cnt = df_sig_resp_final.groupby('cellosaurus_id').size().reset_index(name='counts')

len(df_sig_resp_final['cell_drug'].unique())

df_sig_resp_final

df_sig_resp_final.to_parquet('../../NO_BACKUP/lincs_2020/signature_response_features_r2_top0.7_final.parquet', index=False)

def splitting_by_cell_line(df, 
                           signatures='sig_id', 
                           cell_line='cellosaurus_id', 
                           ratio=0.8,
                           seed=42):
    
    cell_lines = df['cellosaurus_id'].unique()
    train_size = round(ratio * len(cell_lines))
    val_size = round((1 - ratio)/2 * len(cell_lines))
    np.random.seed(seed)
    cell_lines_perm = np.random.permutation(cell_lines)
    train_cell_lines = cell_lines_perm[:train_size]
    val_cell_lines = cell_lines_perm[train_size:train_size + val_size]
    test_cell_lines = cell_lines_perm[train_size + val_size:]
    df_train = df[df['cellosaurus_id'].isin(train_cell_lines)].sample(frac=1).reset_index(drop=True)
    df_val = df[df['cellosaurus_id'].isin(val_cell_lines)].sample(frac=1).reset_index(drop=True)
    df_test = df[df['cellosaurus_id'].isin(test_cell_lines)].sample(frac=1).reset_index(drop=True)
    return df_train, df_val, df_test

df_train, df_val, df_test = splitting_by_cell_line(df_sig_resp_final)

df_train[df_train['cellosaurus_id'].isin(df_test['cellosaurus_id'])]

df_val[df_val['cellosaurus_id'].isin(df_test['cellosaurus_id'])]

import numpy as np
from sklearn.model_selection import KFold

kf = KFold(n_splits=5)
df = df_sig_resp_final.copy()
cell_lines = df['cellosaurus_id'].unique()
cell_lines_perm = np.random.permutation(cell_lines)

for train_index, test_index in kf.split(cell_lines_perm):
    train_cell_lines, test_cell_lines = cell_lines_perm[train_index], cell_lines_perm[test_index]
    df_train = df[df['cellosaurus_id'].isin(train_cell_lines)].sample(frac=1).reset_index(drop=True)
    df_test = df[df['cellosaurus_id'].isin(test_cell_lines)].sample(frac=1).reset_index(drop=True)
    print("train:", cell_lines_perm[train_index], "test:", cell_lines_perm[test_index])
    print("train:", train_index, "test:", test_index)
